Use custom validator for only checking existing factories
