+++ src/Sylius/Bundle/ApiBundle/Resources/config/api_resources/GatewayConfig.xml
@@ -20,7 +20,6 @@
 
         <attribute name="validation_groups">
             <attribute>sylius</attribute>
-            <attribute>sylius_api</attribute>
         </attribute>
 
         <collectionOperations />
