+++ src/Sylius/Bundle/PayumBundle/Resources/config/services.xml
@@ -17,12 +17,7 @@
                                http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <imports>
-        <import resource="services/action.xml" />
-        <import resource="services/controller.xml" />
-        <import resource="services/extension.xml" />
-        <import resource="services/factory.xml" />
-        <import resource="services/form.xml" />
-        <import resource="services/provider.xml" />
+        <import resource="services/**.xml" />
     </imports>
 
     <services>
