+++ src/Sylius/Bundle/PayumBundle/Resources/config/services/validator.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<!--
+
+ This file is part of the Sylius package.
+
+ (c) Sylius Sp. z o.o.
+
+ For the full copyright and license information, please view the LICENSE
+ file that was distributed with this source code.
+
+-->
+
+<container xmlns="http://symfony.com/schema/dic/services"
+           xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+           xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd"
+>
+    <services>
+        <defaults public="true" />
+
+        <service id="Sylius\Bundle\PayumBundle\Validator\GatewayFactoryExistsValidator">
+            <argument>%sylius.gateway_factories%</argument>
+            <tag name="validator.constraint_validator" alias="sylius_gateway_factory_exists_validator" />
+        </service>
+
+        <service id="Sylius\Bundle\PayumBundle\Validator\GroupsGenerator\GatewayConfigGroupsGenerator">
+            <argument>%sylius.payum.gateway_config.validation_groups%</argument>
+        </service>
+    </services>
+</container>
